Enable IAP endpoint all the time
